websitescan/requirements.txt
requests
beautifulsoup4
lxml
python-wappalyzer
certifi
Flask
